+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add transfer from savings button on empty spending wallet when user has on-chain balance #523
+
 ### Changed
 - Use middle-ellipsis truncation for addresses on the receive screen #517